refactor: isolate quantized weight export
