+++ modelopt/torch/export/__init__.py
@@ -21,6 +21,7 @@ from .model_config_export import *
 from .model_utils import *
 from .moe_utils import *
 from .plugins import *
+from .quantized_weight_export import *
 from .registry import *
 from .shard_cast_utils import *
 from .transformer_engine import *
