+++ tests/gpu/torch/export/test_export_weight_gpu.py
@@ -26,17 +26,19 @@ from torch.nn import init
 import modelopt.torch.quantization as mtq
 from modelopt.torch.export.model_config import QUANTIZATION_MXFP8
 from modelopt.torch.export.quant_utils import (
-    build_hf_quantization_config,
-    capture_quantized_weight_export_state,
-    export_quantized_weight_tensors,
     get_activation_scaling_factor,
     get_quantization_format,
     get_weight_block_size,
     get_weight_scaling_factor,
     get_weight_scaling_factor_2,
     postprocess_state_dict,
-    select_quantized_weight_export_state,
     to_quantized_weight,
+)
+from modelopt.torch.export.quantized_weight_export import (
+    build_hf_quantization_config,
+    capture_quantized_weight_export_state,
+    export_quantized_weight_tensors,
+    select_quantized_weight_export_state,
 )
 from modelopt.torch.export.unified_export_hf import _export_quantized_weight
 from modelopt.torch.quantization.nn.modules.quant_module import QuantModule, QuantModuleRegistry
